requirements.txt
Flask==3.1.2
Flask_Login==0.6.3
flask_sqlalchemy==3.1.1
python-dotenv==1.2.1
spotipy==2.25.1
Werkzeug==3.1.3
gunicorn==23.0.0
pytest==7.4.0
pytest-flask==1.3.0
